fix: onPageCreator was considered as a property not method
.
