+++ packages/puppeteer-extra-plugin-session-persistence/package.json
@@ -49,6 +49,7 @@
     "ava": "^2.4.0",
     "npm-run-all": "^4.1.5",
     "puppeteer": "9",
+    "puppeteer-extra": "^3.3.6",
     "rimraf": "^3.0.0",
     "rollup": "^1.27.5",
     "rollup-plugin-commonjs": "^10.1.0",
